requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
awsebcli==3.21.0
blessed==1.20.0
botocore==1.35.20
boto3
celery
cement==2.10.14
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
datasets==2.21.0
dill==0.3.8
dj-database-url
Django==5.1
django-heroku
django-cors-headers==4.4.0
djangorestframework==3.15.2
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
gunicorn==23.0.0
huggingface-hub==0.24.6
idna==3.8
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pathspec==0.10.1
pyarrow==17.0.0
python-dateutil==2.9.0.post0
python-dotenv
pytz==2024.1
psutil==6.0.0
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
redis
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
six==1.16.0
sqlparse==0.5.1
sympy==1.13.2
termcolor==2.4.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.2.0+cpu
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
wcwidth==0.2.13
whitenoise
xxhash==3.5.0
yarl==1.9.11
kombu

-f https://download.pytorch.org/whl/torch_stable.html
